Update to LLVM 9.0.0
This bumps to the latest LLVM release, just for the sake of being up to
date with the latest warnings and code generation improvements. I
checked codegen-test for any obvious regressions in instruction counts
or inlining behavior.

This added some new UBSAN checks. I disabled implicit integer conversion
and truncation, which seems to be an issue in a lot of spots in our code
(particularly StringPiece from gutil), but did fix one issue where it
spotted an actual lossy type truncation from uint64 to uint32.

Change-Id: I50e09bc6f55c193e070e20b5b8ebb8e896fe37b3
Reviewed-on: http://gerrit.cloudera.org:8080/14817
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>