[test] fix HybridClockTest.TestWaitUntilAfter_TestCase2
In case of slow execution (e.g., a TSAN build) and very low estimated
clock error (e.g., 0 as in case of 'system_unsync' clock), calling
HybridClock::WaitUntilAfter() for a fresh past hybrid timestamp returns
Status::OK() instead of Status::TimedOut() regardless of deadline
because the supplied timestamp is seen clearly in the past.

This patch addresses the issue by moving the target timestamp further
into the future.  Now it's 2 milliseconds in the future, but it was
2 microseconds in the future prior to this patch.

Change-Id: I34108359a04b5bd3f2a957faf6d2e9a6dc022eb1
Reviewed-on: http://gerrit.cloudera.org:8080/14823
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>