[test] KUDU-2610 fix flakiness in kudu-admin-test scenario
I saw AdminCliTest.TestSimultaneousLeaderTransferAndAbruptStepdown
failing with timed-out error trying to write data to the test table:

  http://dist-test.cloudera.org/job?job_id=aserbin.1575418627.126650

A log from another occurrence of this issue is attached to
https://issues.apache.org/jira/browse/KUDU-2610

It seems the frequency of the leader change/step-down requests might be
safely decreased to make this test more stable.  The test already have
the frequency of the leader change requests clamped down for ASAN
builds compared with the RELEASE/DEBUG/TSAN case.

Change-Id: I98e792783efa2909d10174f84ddd785f5a968046
Reviewed-on: http://gerrit.cloudera.org:8080/14827
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins