fs: mark 'filesystem_block_size_bytes' optional
While the field is useful when dealing with data directories, if we want
to reuse the path instance files for WAL directories, the field won't
always be necessary. As such, this marks it 'optional' instead of
'required'.

Is this safe? While not strictly backwards compatible as far as protobuf
goes, I'm assuming that the DataDirManager will continue to use the now
optional field. Taken in that context, the switch to optional shouldn't
pose backwards compatibility issues for older clusters. While I can't
guarantee that, I've at least added a comment indicating the appropriate
usage of the field.

Change-Id: I5378391a94e5faa54a11a277d9191d151e0225d5
Reviewed-on: http://gerrit.cloudera.org:8080/14856
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>