KUDU-3005: Reduce size of kudu-python docker image
This patch breaks out the python environment bootstrapping from
the dev environment bootstrapping. This allows us to use the
runtime image for the kudu-python docker image instead of the
giant dev image.

The result of this change is a 440MB uncompressed image instead
of a 1.92GB uncompressed image.

I manually tested the image by running the python example against
a quickstart cluster:
https://github.com/apache/kudu/blob/master/examples/python/basic-python-example/basic_example.py

Change-Id: I159907c4a85e919d98c92df97cd44d0300d1cd29
Reviewed-on: http://gerrit.cloudera.org:8080/14984
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins