[consensus] respond lock-free to RequestVote() if busy
I saw cases of contention on replica's RaftConsensus lock when the
filesystem was slow on updating Raft metadata files to record a vote
that has just been granted by the replica.  As it turned out, the fast
path was also acquiring the RaftConsensus object lock, but it's easy
to avoid that.

This patch updates the code by not acquiring the lock in such cases.
It should help a bit with overflowing of the RaftConsensus RPC queue.

In addition, I turned the LOG(INFO) message about this event into
VLOG(1) since it's not so important to report about such events on the
responder side: its state doesn't change upon responding with
ServiceUnavailable anyways.  Other minor clean-up.

Change-Id: I95d5cbe455fefc4cdc540ee1e7b69e1f21b6ebc0
Reviewed-on: http://gerrit.cloudera.org:8080/14943
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>