Adding a leader detected callback which fires when a follower joins the ring
Summary: with no term change or no no-op received. We keep track if we
also receive a no-op. And only in the case that norcb has not already
been fired, we fire ldcb. We could fire LDCB all the time, but we need
to reconcile on mysql side if there needs to be 2 callbacks or 1 in
future.

Test Plan: Used this to change MASTER TO and start Slave applier on a
follower restart

Reviewers: iRitwik, abhinav04sharma

Subscribers:

Tasks:

Tags: