Maintain ABI compatibility between release and debug builds
Summary: Because the integrations we are currently developing don't
maintain a well defined public API boundary with Kudu, adding fields to
structs in debug mode is unfortunately dangerous and crash prone. At
least for now, we need the ability to build Kudu in release mode and run
linked modules (Raft callbacks, etc) in debug mode. Those modules may
end up #including all kinds of headers (locks, etc) and
difficult-to-understand runtime crashes are the result.

In a perfect world, we would define an ABI-compatible interface inside
of KuduRaft and prevent exposing details of the locks we are using
inside of Log, RaftConsensus, etc. In the current world, we want the
flexibility to hack through and get something working quickly without
having to build in release mode only.

This patch disables several debug-mode-only features in Kudu (keeping
track of mutex owners, etc) because they end up changing the structs
defined in commonly-used header files. This fixes some crashes we've
been seeing when integrating some plugin code.

Test Plan: I ran a gtest that reliably reproduced a nasty crash and now
it doesn't crash anymore.

Reviewers: anirbanr-fb, abhinav04sharma