KUDU-3011 p1: metric to count tablet leaders
Adds a metric that tracks the number of tablet leaders on a given
tablet server. This is done by plumbing down a top-level gauge down from
the KuduServer to each TabletReplica's RaftConsensus instance.

This is done at the KuduServer to eventually be extensible to count the
number of master system catalog leaders, though that plumbing is left
for a future patch. I've left a TODO where I expect this to happen.

I also considered instead having the metric be defined by a functor that
would iterate through all replicas and check each's leadership status. I
opted to not do this, since iterating through and locking each
RaftConsensus instance seemed like it'd be less performant.

This will be useful in orchestrating a smooth maintenance window, as it
will allow us to determine whether leadership has quiesced away from a
given tablet server.

Change-Id: Iaa6554458a860e34f97af168da7ed786c8ef47e4
Reviewed-on: http://gerrit.cloudera.org:8080/14976
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>