test: deflake RaftConsensusNumLeadersMetricTest TestNumLeadersMetric
The test previously wouldn't wait for all tablets to be up and running
on every tserver before doing stuff. This led to errors when attempting
to start elections, since the StartElection RPC would be eaten and never
retried, leaving some tablets with no leader.

The fix is to wait for every tserver to have all running tablets.

Before this fix, the test would fail in TSAN mode 6/1000 times. With it,
it passed 1000/1000.

Change-Id: I4ef4651e4562ac274982773ebb8c46e5137e38b5
Reviewed-on: http://gerrit.cloudera.org:8080/14980
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins