file cache: unify across file types
Back when the file cache was first added (commit 6a124f040), there was a
choice between using a templatized approach, instantiated per-file-type and
a non-templatized approach, usable by any file type. At the time, file cache
usage was restricted to the block managers, and so the templatized approach
carried the day.

Since then we've seen more and more reports of users hitting fd limits due
to high numbers of log segments. I'd like to address these by expanding the
purview of the file cache to include all long-lived opened fds, including
segment files and index chunks.

This patch lays the groundwork for that by de-templatizing the file cache,
"unifying" it for all supported file types. Eventually, it'll be
instantiated once per server and rolled out to the WAL, but for now only the
underlying implementation changes.

Change-Id: I1e5cf6cadb7f0d45492af8b8a7541853d587409c
Reviewed-on: http://gerrit.cloudera.org:8080/14988
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>