maintenance_manager-test: deflake TestPriorityOpLaunch
Sometimes, the test didn't get its desired op execution order because the MM
would execute an op before all of the ops were registered. That is, the ops
weren't registered atomically, and an "unlucky" overlap of FindBestOp and
one of the registration calls would lead to an errant op execution.

We can fix this by using FLAGS_enable_maintenance_manager to disable the MM
while we register ops.

Before the fix, I could reliably reproduce a failure within 1000 executions
of the test.

After, I couldn't reproduce it in 30000 executions.

Change-Id: I06f3fd8fa952ed74bfaf2ac3dd55cd289f4b878b
Reviewed-on: http://gerrit.cloudera.org:8080/15025
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>