KUDU-3042: Fix invalid DCHECK when rpc is cancelled then times out
The timeout handling code in Connection::HandleOutboundCallTimeout
assumes that when a timeout is triggered the OutboundCall must still
be valid. In the case of an asynchronous rpc that was cancelled before
the timeout was hit, this will not be the case.

The fix is to check if the OutboundCall is still valid and return
without processing the timeout if it is not.

Testing:
- Added timeout tests to TestCancellation to inject timeouts at
  OutboundCall states from READY to SENT.
- Tested in Impala in the context of implementing IMPALA-8712.

Change-Id: I5b97ccc4787809fca8fbc991bb1151940477efca
Reviewed-on: http://gerrit.cloudera.org:8080/15032
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>