webserver: use faster gzip compression
If an HTTP client indicates that it supports gzip, we gzip-compress the
web response. This can be beneficial over slow links, but the default
gzip compression level is pretty slow. This changes the gzip compression
to use the fastest level, which is several times faster, without that
much loss in space.

Change-Id: If97d883483a7313bd5496af68b17e968a5223b1a
Reviewed-on: http://gerrit.cloudera.org:8080/15059
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>