[gutil] Rename ARRAYSIZE() to KUDU_ARRAYSIZE()
Summary: The ARRAYSIZE() macro is commonly written across codebases, and
the one Kudu uses from gtuil conflicts with another library I want to
link kuduraft with. The simplest way for me to resolve this conflict is
to namespace the Kudu version of the macro.

Test Plan: `cmake ../.. && make`

Reviewers: anirbanr-fb, iRitwik

Subscribers:

Tasks: T60456824

Tags: