When we are doing a first run, assumption is we are creating a fresh raft instance.
Summary: In that case, the recovery does not have to get the OpIds
from the binlog files. It is also likely that first run has no
opids because we are moving from non-raft to raft mode.

Test Plan: This was done to prevent wrong behavior when we wanted to
create a new raft ring. Log::Init in myshadow would go through all the
previous binlog files and look for Last OpId.

Reviewers: mpercy

Subscribers:

Tasks:

Tags: