[clock] auto-config of built-in NTP client in cloud
This patch introduces auto-configuration of the built-in NTP client
in public cloud environment.  Currently, AWS and GCE public cloud types
are supported: Kudu masters and tablet servers are now capable of
auto-detecting per-instance NTP server and using it as reference server
for the built-in NTP client.

The auto-configuration is controlled by the
--builtin_ntp_client_enable_auto_config_in_cloud boolean flag and gated
by the --time_source flag (i.e. the latter should be set to 'builtin'
to allow the auto-configuration to work).

Change-Id: I0590c0b731a4da2f968e720dea0410d46ab62beb
Reviewed-on: http://gerrit.cloudera.org:8080/15070
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>