[master] Replace hive_metastore_sasl_enabled validator
The hive_metastore_sasl_enabled GROUP_FLAG_VALIDATOR
that was added in a76177f was accidentally removed in 4f82a46.

The validation is added back in this patch. However, it is
validated in RunMasterServer() as opposed to using a
GROUP_FLAG_VALIDATOR because the master server can be run
from the tools now.

Change-Id: I9ecea5b4b4e0d4bdd198241f23bf34ea46ff6945
Reviewed-on: http://gerrit.cloudera.org:8080/15076
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>