[tests] minor clean up on linked_list test
This patch updates the linked_list test to output more information
in case if the curl fetcher thread fails to retrieve data from one
of the monitored URLs.

I also updated the code to handle non-OK status returned from
KuduScanner::GetCurrentServer(), replaced gscoped_ptr with
std::unique_ptr, and did other minor updates.

The motivation for this change is get more details on the error
if it happens.

Change-Id: Ic92c3a41cb3f37426c32a2c61a11c400cf855cf7
Reviewed-on: http://gerrit.cloudera.org:8080/15100
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>