[tests] fix flakiness in RaftConsensusNonVoterITest
Fixed flakiness in PromoteAndDemote scenario of the
RaftConsensusNonVoterITest test.

Due to the short heartbeat interval and concurrent activity, sometimes
it takes a bit more time for a tserver to respond to
RequestConsensusVote call than the Raft timeout.

Prior to this fix, the scenario failed for ASAN builds in rate of
1 in ~100 runs.  With the fix, I could not see it failing after
multiple 1K runs.

Change-Id: Ia76e4ea57f2f9d4251ad3cbed47267266fee7e28
Reviewed-on: http://gerrit.cloudera.org:8080/15115
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>