[net_util] Fix parsing of IPv6 addr:port strings
In commit ad6a7c3115cd8bf4413cfcb2cdf1be457a6008db, support for IPv6
addr:port parsing was added, however the following edge cases were not
handled:

1. Correctly parse an IPv6 address with no port, i.e. "[::1]"
2. Return an error when a port specifier exists but no port was given,
   i.e. "[::1]:"
3. Correctly detect the wildcard address, "::"

This patch also updates the IPv4 tests to work on (and actually require)
IPv6. To upstream this patch, the tests would need to be modified to
work for both IP versions.

Note: Not handled in this patch: Specifying an IPv6 address (with or
without a port) without enclosing square brackets, even though we don't
enclose them in brackets when stringifying them, making that logic
asymmetric.