KUDU-2780: create thread for auto-rebalancing
The auto-rebalancer thread is a background task of
the master's catalog manager. Each loop iteration, the
thread collects information on tablet servers, tables,
and tablets, in order to determine the best rebalancing
replica moves for the current state of the cluster.
The maximum number of moves per tserver, per iteration,
is controlled by a flag. Currently, each batch of moves
scheduled per iteration is executed synchronously. The
thread waits for the moves to complete, then sleeps for a
time interval, which is controlled by a flag, before resuming.

If the cluster has placement policy violations, the thread
will prioritize scheduling and executing replica moves to
reinstate the policy. Otherwise, the thread will perform
inter-location (cross-location), then intra-location (by
table, then by tserver) rebalancing.

By default, auto-rebalancing is disabled. This can be
changed by a flag in the catalog manager. The rebalancer
tool should be run first, to fully rebalance the cluster,
before enabling auto-rebalancing.

Change-Id: Ifca25d1063c07047cf2123e6792b3c7395be20e4
Reviewed-on: http://gerrit.cloudera.org:8080/14177
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>