[java] Document increasing Intellij intellisense configuration
Our Protobuf files generate Java classes that are too large for
Intellij to provide code assistance for by default.
This results in what appears to be missing classes and difficulty
working with the Protobuf classes.

I added a note under the Intellij section of the Java readme to help
others workaround this.

Change-Id: I37ed4fd9b838fb363a91aa08acae8e7dbb13939a
Reviewed-on: http://gerrit.cloudera.org:8080/15155
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>