[util] remove timeout parameter for cloud::InstanceDetector
With this patch, the detector relies solely on the timeouts reported
by metadata fetchers and doesn't set the timeout on the auto-detection
process itself.  This change simplifies the semantics of the cloud
instance detection.

In addition, and this patch fixes intermittent failures recently seen
in the following test scenarios when running on non-cloud hosts:
  * InstanceDetectorTest.Basic
  * BuiltinNtpWithMiniChronydTest.CloudInstanceNtpServer

Change-Id: I6f7cac9273e884abda22b9e9a7d37fa3e307c2d4
Reviewed-on: http://gerrit.cloudera.org:8080/15168
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>