KUDU-2973 Add semi-database support for Ranger
The table names are "normalized" if HMS integration is enabled so that
they can be synchronized to HMS. Hive's table/database identifiers are
pretty restrictive (alphanumerical ASCII characters + underscore and
forward slash) and it's case-insensitive. As Kudu doesn't support
databases it imposes a further restriction: the table identifiers to be
synchronized with HMS have to contain exactly one "." character which
separates the database name from the table name in HMS.

Ranger's restrictions regarding identifiers are more lax, can be case
sensitive (will not be used for Kudu) and is not limited to ASCII.

This commit introduces a method to parse database and table name for
Ranger from a table identifier that works slightly different from Hive's
table name parser, as it allows non-ASCII characters and provides a
configurable default database name when the table name doesn't contain a
period.

Change-Id: I11431ff5bc75540edff56ef3d4ad384fa37d33d5
Reviewed-on: http://gerrit.cloudera.org:8080/15018
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>